+++ app/src/main/java/de/levithas/aixdroid/domain/model/ModelData.kt
@@ -10,7 +10,7 @@ data class ModelData(
     val author: String = "",
     val licence: String = "",
 
-    val timePeriod: Float = 1.0f,
+    val timePeriod: Int = 1,
     val n_steps: Int = 1,
 
     val inputs: List<TensorData> = emptyList(),
